removing sdk dep and renaming page component
