+++ examples/next-app-directory/package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "@builder.io/react": "^2.0.15",
-    "@builder.io/sdk": "^1.1.33",
     "@builder.io/widgets": "^1.2.24",
     "@types/node": "18.14.6",
     "@types/react": "18.0.28",
