+++ examples/next-app-directory/src/app/[[...slug]]/BuilderPage.tsx
